Update command line options, refactor main()
The timestamp save was given better command line flags, 's' to
autosave the timestamp, and 'S' to prevent the save. Long options
are --save-timestamp (the default) and --skip-timestamp.

A new command line option that allows an RSS feed to be added for
download, instead of using the default feed.
    usage: showrss -f 'http://path/to/feed.rss'

main() was also refactored to simplify the call to main(args), and
this was also when all timestamp activity was moved into main().